+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Property
     - Mandatory field for the rent index url
+    - Mandatory field for the rent increase capping limit
 - Rent increase
     - Generate notification document on rent increase